refactor(frontend): group entity links under entity-links/
Move UserLink and LocationLink from the components top level into a
small family folder. Both are route-aware entity links: ui/ is out, and
LocationLink's cross-route use means it must live in $lib, where a lone
single-occupant folder would be awkward.
