+++ frontend/src/lib/components/entity-links/LocationLink.svelte
