+++ frontend/src/lib/components/entity-links/UserLink.svelte
